# splinterd/api/static/openapi.yml
...
openapi: '3.0.0'
info:
  version: 0.3.12
  title: SplinterD API
  description: REST API for SplinterD
servers:
  - url: http://localhost:9000/api
paths:
...
  /scabbard/{circuit}/{service_id}/batches:
    post:
      description: Send a list of Sabre batches to the specified Scabbard service
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: circuit
          in: path
          description: circuit the targeted service belongs to
          required: true
          schema:
            type: string
        - name: service_id
          in: path
          description: ID of the targeted service
          required: true
          schema:
            type: string
      responses:
        202:
          description: Batch has been submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        404:
          description: The scabbard service was not found, or the circuit id or service id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scabbard/{circuit}/{service_id}/batch_statuses:
    post:
      description: Send a list of Sabre batches to the specified Scabbard service
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: circuit
          in: path
          description: circuit the targeted service belongs to
          required: true
          schema:
            type: string
        - name: service_id
          in: path
          description: ID of the targeted service
          required: true
          schema:
            type: string
        - name: ids
          in: query
          description: Comma separated list of batch ids
          required: true
          schema:
            type: string
        - name: wait
          in: query
          description: time (in seconds) to wait for batches to be committed
          required: false
          schema:
            type: integer
            default: 300
      responses:
        200:
          description: Batch has been submitted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchStatus'
        404:
          description: The scabbard service was not found, or the circuit id or service id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        400:
          description: Malformed ids list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        408:
          description: Batches did not complete before specified wait
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scabbard/{circuit}/{service_id}/state:
    get:
      description: Experimental - Get a list of entries in state from the specified Scabbard service
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: circuit
          in: path
          description: circuit the targeted service belongs to
          required: true
          schema:
            type: string
        - name: service_id
          in: path
          description: ID of the targeted service
          required: true
          schema:
            type: string
        - name: prefix
          in: query
          description: The address prefix for filtering state entries
          required: false
          schema:
            type: string
            example: 00ec01
      responses:
        200:
          description: List of entries in state (under given address prefix, if specified)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    address:
                      type: string
                    value:
                      type: array
                      items:
                        type: integer
        400:
          description: Request was malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: The scabbard service was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scabbard/{circuit}/{service_id}/state/{address}:
    get:
      description: Experimental - Get the value at a given address in state from the specified Scabbard service
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: circuit
          in: path
          description: circuit the targeted service belongs to
          required: true
          schema:
            type: string
        - name: service_id
          in: path
          description: ID of the targeted service
          required: true
          schema:
            type: string
        - name: address
          in: path
          description: The address of the value to retrieve from state
          required: true
          schema:
            type: string
            example: 000000a87cb5eafdcca6a814e4add97c4b517d3c530c2f44b31d18e3b0c44298fc1c14
      responses:
        200:
          description: The value at the requested address
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
        400:
          description: Request was malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: The scabbard service was not found, or there is no value at the given address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    protocol_version:
      schema:
        type: integer
        example: 1
      description: The protocol version which the client can understand
      in: header
      name: SplinterProtocolVersion
  schemas:
    Error:
      additionalProperties: false
      properties:
        message:
          description: A message describing the error that occurred
          type: string
          example: DatabaseError({description})
      required:
        - message
...
    Link:
      type: object
      properties:
        link:
          type: string
          description: link to batch_statuses endpoint
          example: "/scabbard/{circuit_id}/{service_id}/batch_statuses?ids={ids}"
    BatchStatus:
      type: object
      properties:
        id:
          type: string
          description: Batch header signature
          example: 6ff35474a572087e08fd6a54d563bd8172951b363e5c9731f1a40a855e14bba45dac515364a08d8403f4fb5d4a206174b7f63c29e4f4e425dc71b95494b8a798
        status:
          type: object
          description: Batch status
          properties:
            statusType:
              type: string
              enum:
                - Unknown
                - Invalid
                - Valid
                - Pending
                - Committed
            message:
                type: array
                items:
                  type: object
                  properties:
                    transaction_id:
                      type: string
                      example: f4e147ff464013deccb3e68bb8619beffb29ff86b401257c93bcf8ef76d7ca173fa84b4f4a58414ad2d00a2c9f810cbb726e01cd26ebd44720239d9d35853099
                    error_message:
                      type: string
                      example: "Wasm contract returned invalid transaction: xo, 0.3.3"
                    error_data:
                      type: array
                      items:
                        type: integer
...